requirements.txt
PyQt5>=5.15.0
pyotp>=2.8.0
qrcode>=7.3.0
Pillow>=8.0.0
pyzbar>=0.1.9
pycryptodome>=3.15.0